requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
cloudinary==1.33.0
Django==4.1.1
django-cloudinary-storage==0.3.0
djangocodemirror==2.1.0
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.4
numpy==1.25.2
openai==0.24.0
openpyxl==3.1.2
pandas==2.0.3
pandas-stubs==2.0.2.230605
Pillow==9.1.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2023.3
requests==2.31.0
six==1.16.0
SQLAlchemy==1.2.19
sqlparse==0.4.4
tqdm==4.65.0
types-pytz==2023.3.0.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
whitenoise==6.2.0
